requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.4.5
djangorestframework==3.13.1
numpy==1.23.1
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
